fix: testnet review F-2..F-6 + redeploy all 3 testnets (2026-09-24/25)
- FreelancerEscrow: rejected-milestone resubmit (F-2), hireGig stale-deadline re-check (F-4)
- FeeDistributor: setSplitBps accepts sum <= BPS (F-3)
- UserProfile: uint32 review counters (F-5), resolveDispute decrements counter (F-6)
- Deploy scripts: deploy_all gains post-deploy config + manualPrice; new resume_deploy_amoy
  (checkpointed/resumable per best-deploy-contract.md), configure_deployed, redeploy_fe
- Redeployed + verified 7/7 on Base Sepolia & Arbitrum Sepolia (09-24) and Polygon Amoy (09-25)
- Docs: README/SECURITY/contracts README refreshed with new address tables + F-review table,
  dead Amoy RPC replaced with drpc/publicnode, stale checkpoint untracked (gitignored)
- RWA: manual price typo fixed (.00 -> $0.20 POL/USD) on-chain and in scripts
- Tests: 144 Hardhat + 15 Foundry passing; next build clean

+++ frontend/src/hooks/useContracts.ts
@@ -64,7 +64,7 @@ const USER_PROFILE_ABI = [
   { inputs: [{ name: "user", type: "address" }], name: "reviewScore", outputs: [{ name: "", type: "uint8" }], stateMutability: "view", type: "function" },
   { inputs: [{ name: "user", type: "address" }], name: "compositeScore", outputs: [{ name: "", type: "uint8" }], stateMutability: "view", type: "function" },
   { inputs: [{ name: "user", type: "address" }], name: "getProgress", outputs: [{ name: "", type: "uint8" }], stateMutability: "view", type: "function" },
-  { inputs: [{ name: "user", type: "address" }], name: "getReviewBreakdown", outputs: [{ name: "pos", type: "uint8" }, { name: "neg", type: "uint8" }], stateMutability: "view", type: "function" },
+  { inputs: [{ name: "user", type: "address" }], name: "getReviewBreakdown", outputs: [{ name: "pos", type: "uint32" }, { name: "neg", type: "uint32" }], stateMutability: "view", type: "function" },
   { inputs: [{ name: "user", type: "address" }], name: "passportScore", outputs: [{ name: "", type: "uint8" }], stateMutability: "view", type: "function" },
   { inputs: [{ name: "user", type: "address" }], name: "biometricVerified", outputs: [{ name: "", type: "bool" }], stateMutability: "view", type: "function" },
   { inputs: [{ name: "user", type: "address" }], name: "thirdPartyScore", outputs: [{ name: "", type: "uint8" }], stateMutability: "view", type: "function" },
